create SpotifyService and implemented oauth authentication and getSong
Credentials for the application are loaded from a file that is not
checked into version control for security reasons.
ensureAccessToken() is a method that gets the access token from spotify
if the server doesn't currently have a valid one.
Each method that makes calls to Spotify should call ensureAccessToken
first.