implement onSongSelected and add Question class
onSongSelected takes the songid selected by player and orchestrates
calls to SpotifyService to get the selected song and its similar songs.

Added a Question class to encapsulate the song options and which
song is 'correct'. onSongSelected instantiates and stores the current
Question object as an instance variable of the Game to be used later
when players answer the question.