implement onAnswer in Game and destroy game event and handlers
Implemented onAnswer in Game to check if player answers are correct
and record point increases for players. Added logic for determining
if a player has won and passing the appropriate state information
to PlayerState.

Also added an EventEmitter in ServerController that gets passed to
all Game objects it creates so that the games can tell the
ServerController
when to destroy it.

Implemented destroy functions in Game, PlayerState, PlayerController
to properly destroy the objects and their members if required.