design changes to PlayerState and Game handlers
Refactored PlayerState and Game so that PlayerState keeps track of
a "madeMove" flag which indicates whether the player has already
made their input/move for the current state.

Game then uses this madeMove flag in its handlers for player actions
whenever it needs to check if all players have provided input so that
the game's state can be advanced.

Also refactored the sending of events to all players in a given Game
to be done from ServerController instead of through a single
PlayerController.
Whenever Game wants to send an event to all players in it, it can
use the EventEmitter it shares with ServerController and pass it its
id, name of the event, and data.